show confirmation and loading messages - sweetAlert2

+++ src/modules/daybook/views/EntryView.vue
@@ -51,7 +51,9 @@
 
 <script>
 import { defineAsyncComponent } from 'vue'
-import { mapActions, mapGetters } from "vuex"
+import { mapActions, mapGetters } from 'vuex'
+import Swal from 'sweetalert2'
+
 import getDayMonthYear from '../helpers/getDayMonthYear'
 
 export default {
@@ -86,20 +88,47 @@ export default {
             this.entry = entry
         },
         async saveEntry(){
+
+            new Swal({
+                title: 'Espere por favor',
+                allowOutsideClick: false
+            })
+            Swal.showLoading()
+
             if(this.entry.id){
-                //Update
+                //Update entry
                 await this.updateEntry(this.entry)
             }
             else{
-                //Add
+                //Add new entry
                 const id = await this.createEntry(this.entry)
                 this.$router.push({name: 'entry', params:{id}})
             }
 
+            Swal.fire('Guardado', 'Entrada registrada con éxito', 'success')
+
         },
         async onDeleteEntry(){
-            await this.deleteEntry(this.entry.id)
-            this.$router.push({name: 'no-entry'})
+
+            const {isConfirmed} = await Swal.fire({
+                title: '¿Estás seguro?',
+                text: 'Una vez borrado no se puede recuperar',
+                showDenyButton: true,
+                confirmButtonText: 'Sí, estoy seguro'
+            })
+
+            if(isConfirmed){
+                new Swal({
+                    title: 'Espere por favor',
+                    allowOutsideClick: false
+                })
+                Swal.showLoading()
+                await this.deleteEntry(this.entry.id)
+                this.$router.push({name: 'no-entry'})
+
+                Swal.fire('Eliminado','', 'success')
+            }
+
         },
         ...mapActions('journal', ['updateEntry', 'createEntry', 'deleteEntry'])
     },
